Revert "Enhancement/navigation component (#9)"
This reverts commit c7e863c13ff2b96ddbffc5b355ea708ec7a8f7a5.
