+++ app/build.gradle
@@ -10,8 +10,6 @@ apply plugin: 'com.google.gms.google-services'
 
 apply plugin: 'com.google.firebase.crashlytics'
 
-apply plugin: "androidx.navigation.safeargs.kotlin"
-
 android {
     compileSdkVersion 29
 
@@ -127,8 +125,6 @@ android {
 }
 
 dependencies {
-    def nav_version = "2.3.3"
-
     implementation fileTree(include: ['*.jar'], dir: 'libs')
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
     implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-android:1.3.2'
@@ -186,12 +182,6 @@ dependencies {
     //---- Location
     implementation "com.google.android.gms:play-services-location:17.1.0"
     implementation 'com.google.android.gms:play-services-basement:17.5.0'
-
-    //--Navigation
-    implementation "androidx.navigation:navigation-fragment-ktx:$nav_version"
-    implementation "androidx.navigation:navigation-ui-ktx:$nav_version"
-    implementation "androidx.navigation:navigation-dynamic-features-fragment:$nav_version"
-    implementation "androidx.navigation:navigation-compose:1.0.0-alpha06"
 
     //---- Linters ----
     ktlint "com.pinterest:ktlint:0.35.0"
