+++ app/src/androidTest/java/com/rootstrap/android/tests/ProfileActivityTest.kt
@@ -6,8 +6,8 @@ import androidx.test.espresso.action.ViewActions.click
 import androidx.test.espresso.matcher.ViewMatchers.withId
 import com.rootstrap.android.R
 import com.rootstrap.android.network.managers.SessionManager
-import com.rootstrap.android.ui.activity.main.authentication.AuthenticationActivity
 import com.rootstrap.android.ui.activity.main.authentication.ProfileActivity
+import com.rootstrap.android.ui.activity.main.authentication.SignUpActivity
 import com.rootstrap.android.utils.BaseTests
 import okhttp3.mockwebserver.Dispatcher
 import okhttp3.mockwebserver.MockResponse
@@ -43,7 +43,7 @@ class ProfileActivityTest : BaseTests() {
         // Check if this activity was successful launched
         activity.runOnUiThread {
             val current = currentActivity()
-            assertEquals(AuthenticationActivity::class.java.name, current::class.java.name)
+            assertEquals(SignUpActivity::class.java.name, current::class.java.name)
         }
     }
