+++ app/src/androidTest/java/com/rootstrap/android/tests/SignUpActivityTest.kt
@@ -1,5 +1,6 @@
 package com.rootstrap.android.tests
 
+import androidx.test.core.app.ActivityScenario
 import androidx.test.espresso.Espresso
 import androidx.test.espresso.action.ViewActions
 import androidx.test.espresso.action.ViewActions.replaceText
@@ -8,6 +9,9 @@ import com.google.gson.Gson
 import com.rootstrap.android.R
 import com.rootstrap.android.network.managers.SessionManager
 import com.rootstrap.android.network.models.UserSerializer
+import com.rootstrap.android.ui.activity.main.authentication.ProfileActivity
+import com.rootstrap.android.ui.activity.main.authentication.SignInActivity
+import com.rootstrap.android.ui.activity.main.authentication.SignUpActivity
 import com.rootstrap.android.utils.BaseTests
 import okhttp3.mockwebserver.Dispatcher
 import okhttp3.mockwebserver.MockResponse
@@ -17,22 +21,21 @@ import org.junit.Assert.assertEquals
 import org.junit.Before
 import org.junit.Test
 
-// TODO:migrate test to the fragment
 class SignUpActivityTest : BaseTests() {
 
-    // private lateinit var activity: SignUpActivity
-    // private lateinit var scenario: ActivityScenario<SignUpActivity>
+    private lateinit var activity: SignUpActivity
+    private lateinit var scenario: ActivityScenario<SignUpActivity>
 
     @Before
     override fun before() {
         super.before()
-        //   scenario = FragmentScenario.launch(SignUpFragment::class.java)
-        //   scenario.onActivity { activity -> this.activity = activity }
+        scenario = ActivityScenario.launch(SignUpActivity::class.java)
+        scenario.onActivity { activity -> this.activity = activity }
     }
 
     @Test
     fun signUpSuccessfulTest() {
-        //  scenario.recreate()
+        scenario.recreate()
         setServerDispatch(signUpDispatcher())
         val testUser = testUser()
         scrollAndTypeText(R.id.first_name_edit_text, testUser.firstName)
@@ -42,27 +45,25 @@ class SignUpActivityTest : BaseTests() {
 
         Espresso.onView(ViewMatchers.withId(R.id.gender_drop_down_text)).perform(
             ViewActions.scrollTo(),
-            replaceText(testUser.gender)
-        )
+            replaceText(testUser.gender))
 
         scrollAndPerformClick(R.id.sign_up_button)
         val user = SessionManager.user
         assertEquals(user, testUser)
-        /* activity.runOnUiThread {
-             val current = currentActivity()
-             assertEquals(ProfileActivity::class.java.name, current::class.java.name)
-         } */
+        activity.runOnUiThread {
+            val current = currentActivity()
+            assertEquals(ProfileActivity::class.java.name, current::class.java.name)
+        }
     }
 
     @Test
     fun checkCTASignIn() {
-        //   scenario.recreate()
+        scenario.recreate()
         scrollAndPerformClick(R.id.sign_in_text_view)
-        /*   activity.runOnUiThread {
-               val current = currentActivity()
-               assertEquals(SignInActivity::class.java.name, current::class.java.name)
-           }
-           */
+        activity.runOnUiThread {
+            val current = currentActivity()
+            assertEquals(SignInActivity::class.java.name, current::class.java.name)
+        }
     }
 
     private fun signUpDispatcher(): Dispatcher {
