+++ app/src/main/AndroidManifest.xml
@@ -5,8 +5,8 @@
     <uses-permission android:name="android.permission.ACCESS_FINE_LOCATION" />
     <uses-permission android:name="android.permission.INTERNET" />
     <uses-permission android:name="android.permission.CAMERA" />
-    <uses-permission android:name="android.permission.ACCESS_COARSE_LOCATION" />
-    <uses-permission android:name="android.permission.ACCESS_FINE_LOCATION" />
+    <uses-permission android:name="android.permission.ACCESS_COARSE_LOCATION"/>
+    <uses-permission android:name="android.permission.ACCESS_FINE_LOCATION"/>
 
     <application
         android:name=".App"
@@ -18,15 +18,6 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
 
-        <activity android:name=".ui.activity.main.splash.SplashActivity"
-            android:screenOrientation="portrait">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
-        </activity>
-
         <meta-data
             android:name="com.google.android.geo.API_KEY"
             android:value="@string/google_maps_key" />
@@ -35,14 +26,20 @@
             android:name=".ui.activity.main.targetpoint.TargetPointsActivity"
             android:label="@string/title_activity_map"
             android:screenOrientation="portrait" />
-
+        <activity
+            android:name=".ui.activity.main.authentication.SignUpActivity"
+            android:screenOrientation="portrait" />
         <activity
             android:name=".ui.activity.main.authentication.ProfileActivity"
             android:screenOrientation="portrait" />
         <activity
-            android:name=".ui.activity.main.authentication.AuthenticationActivity"
+            android:name=".ui.activity.main.authentication.SignInActivity"
             android:screenOrientation="portrait">
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
 
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
         </activity>
 
         <meta-data
@@ -66,5 +63,4 @@
             </intent-filter>
         </activity>
     </application>
-
 </manifest>
