--- app/src/main/java/com/rootstrap/android/ui/activity/main/authentication/AuthenticationActivity.kt
@@ -1,24 +1,0 @@
-package com.rootstrap.android.ui.activity.main.authentication
-
-import android.os.Bundle
-import androidx.navigation.findNavController
-import com.rootstrap.android.R
-import com.rootstrap.android.databinding.ActivitySignInBinding
-import com.rootstrap.android.metrics.Analytics
-import com.rootstrap.android.metrics.PageEvents
-import com.rootstrap.android.metrics.VISIT_SIGN_IN
-import com.rootstrap.android.ui.base.BaseActivity
-
-class AuthenticationActivity : BaseActivity() {
-
-    private lateinit var binding: ActivitySignInBinding
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        Analytics.track(PageEvents.visit(VISIT_SIGN_IN))
-        binding = ActivitySignInBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-    }
-
-    override fun onSupportNavigateUp() = findNavController(R.id.nav_host_fragment_container).navigateUp()
-}
