+++ app/src/main/java/com/rootstrap/android/ui/activity/main/authentication/ProfileActivity.kt
@@ -35,7 +35,7 @@ class ProfileActivity : BaseActivity(), ProfileView {
     }
 
     override fun goToFirstScreen() {
-        startActivityClearTask(AuthenticationActivity())
+        startActivityClearTask(SignInActivity())
     }
 
     // ViewModelListener
