+++ app/src/main/java/com/rootstrap/android/ui/activity/main/authentication/SignInActivity.kt
@@ -2,11 +2,7 @@ package com.rootstrap.android.ui.activity.main.authentication
 
 import android.content.Intent
 import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
 import androidx.lifecycle.ViewModelProvider
-import androidx.navigation.findNavController
 import com.facebook.AccessToken
 import com.facebook.CallbackManager
 import com.facebook.FacebookCallback
@@ -14,46 +10,44 @@ import com.facebook.FacebookException
 import com.facebook.login.LoginManager
 import com.facebook.login.LoginResult
 import com.rootstrap.android.R
-import com.rootstrap.android.databinding.FragmentSignInBinding
+import com.rootstrap.android.databinding.ActivitySignInBinding
+import com.rootstrap.android.metrics.Analytics
+import com.rootstrap.android.metrics.PageEvents
+import com.rootstrap.android.metrics.VISIT_SIGN_IN
 import com.rootstrap.android.network.models.User
-import com.rootstrap.android.ui.base.BaseFragment
+import com.rootstrap.android.ui.activity.main.targetpoint.TargetPointsActivity
 import com.rootstrap.android.ui.view.AuthView
 import com.rootstrap.android.util.NetworkState
 import com.rootstrap.android.util.ViewModelListener
 import com.rootstrap.android.util.extensions.value
+import com.rootstrap.android.util.permissions.PermissionActivity
 
-class SignInFragment : BaseFragment(), AuthView {
+class SignInActivity : PermissionActivity(), AuthView {
 
     private lateinit var viewModel: SignInActivityViewModel
 
-    private lateinit var binding: FragmentSignInBinding
+    private lateinit var binding: ActivitySignInBinding
 
     private lateinit var faceBookCallbackManager: CallbackManager
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        arguments?.let {}
-    }
+        binding = ActivitySignInBinding.inflate(layoutInflater)
 
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        binding = FragmentSignInBinding.inflate(layoutInflater, container, false)
+        setContentView(binding.root)
 
-        initView()
+        Analytics.track(PageEvents.visit(VISIT_SIGN_IN))
 
         val factory = SignInActivityViewModelFactory(viewModelListener)
 
-        viewModel = ViewModelProvider(requireActivity(), factory)
+        viewModel = ViewModelProvider(this, factory)
             .get(SignInActivityViewModel::class.java)
 
         logInCallbackFacebook()
 
-        lifecycle.addObserver(viewModel)
+        initView()
 
-        return binding.root
+        lifecycle.addObserver(viewModel)
     }
 
     private fun initView() {
@@ -63,7 +57,11 @@ class SignInFragment : BaseFragment(), AuthView {
     }
 
     override fun showMainPage() {
-        binding.root.findNavController().navigate(R.id.targetPointsActivity)
+        startActivityClearTask(TargetPointsActivity())
+    }
+
+    private fun logInWithFacebook() {
+        LoginManager.getInstance().logIn(this, arrayListOf(FACEBOOK_PERMISSION))
     }
 
     private fun logInCallbackFacebook() {
@@ -94,12 +92,8 @@ class SignInFragment : BaseFragment(), AuthView {
         }
     }
 
-    private fun logInWithFacebook() {
-        LoginManager.getInstance().logIn(this, arrayListOf(FACEBOOK_PERMISSION))
-    }
-
     private fun goToSignUp() {
-        binding.root.findNavController().navigate(R.id.action_signInFragment_to_signUpFragment)
+        startActivity(Intent(this, SignUpActivity::class.java))
     }
 
     private fun signIn() {
