+++ app/src/main/java/com/rootstrap/android/ui/activity/main/targetpoint/TargetPointsActivity.kt
@@ -50,6 +50,7 @@ class TargetPointsActivity : BaseActivity() {
         })
     }
 
+    // TODO: show target in map and remove toast
     private fun successCreatingTarget() {
         createTargetView.expandCollapseCreateTargetSheet()
         hideKeyboard()
