+++ app/src/main/java/com/rootstrap/android/ui/base/BaseActivity.kt
@@ -2,23 +2,47 @@ package com.rootstrap.android.ui.base
 
 import android.annotation.SuppressLint
 import android.app.Activity
+import android.app.AlertDialog
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
-import com.rootstrap.android.util.DialogUtil
+import com.rootstrap.android.R
+import com.rootstrap.android.ui.custom.LoadingDialog
 
 @SuppressLint("Registered")
 open class BaseActivity : AppCompatActivity(), BaseView {
 
+    private var loadingDialog: LoadingDialog? = null
+
     override fun showProgress() {
-        DialogUtil.showProgress(this)
+        if (loadingDialog == null) {
+            loadingDialog = LoadingDialog(this, null)
+        }
+
+        loadingDialog!!.show()
     }
 
     override fun hideProgress() {
-        DialogUtil.hideProgress()
+        if (loadingDialog != null) {
+            loadingDialog!!.dismiss()
+        }
     }
 
     override fun showError(message: String?) {
-        DialogUtil.showError(this, message)
+        val builder = AlertDialog.Builder(this)
+        builder.setTitle(getString(R.string.error))
+
+        when (message) {
+            "" -> builder.setMessage(getString(R.string.generic_error))
+            null -> builder.setMessage(getString(R.string.generic_error))
+            else -> builder.setMessage(message)
+        }
+
+        builder.setPositiveButton(getString(R.string.ok)) { dialog, _ ->
+            dialog.cancel()
+        }
+
+        val dialog: AlertDialog = builder.create()
+        dialog.show()
     }
 
     protected fun startActivityClearTask(activity: Activity) {
