+++ app/src/main/java/com/rootstrap/android/ui/base/BaseFragment.kt
@@ -1,19 +1,18 @@
 package com.rootstrap.android.ui.base
 
 import androidx.fragment.app.Fragment
-import com.rootstrap.android.util.DialogUtil
 
 open class BaseFragment : Fragment(), BaseView {
 
     override fun showProgress() {
-        DialogUtil.showProgress(requireContext())
+        TODO("not implemented") // To change body of created functions use File | Settings | File Templates.
     }
 
     override fun hideProgress() {
-        DialogUtil.hideProgress()
+        TODO("not implemented") // To change body of created functions use File | Settings | File Templates.
     }
 
     override fun showError(message: String?) {
-        DialogUtil.showError(requireContext(), message)
+        TODO("not implemented") // To change body of created functions use File | Settings | File Templates.
     }
 }
