--- app/src/main/res/anim/slide_down.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<set xmlns:android="http://schemas.android.com/apk/res/android" >
-
-    <translate
-        android:duration="@android:integer/config_longAnimTime"
-        android:fromYDelta="0"
-        android:toYDelta="100%" />
-</set>
