+++ app/src/main/res/layout/activity_sign_up.xml
@@ -8,7 +8,8 @@
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
-        android:layout_height="wrap_content">
+        android:layout_height="wrap_content"
+        tools:context=".ui.activity.main.authentication.SignUpActivity">
 
         <FrameLayout
             android:id="@+id/frame_header"
