--- app/src/main/res/layout/activity_splash.xml
@@ -1,28 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.motion.widget.MotionLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:id="@+id/splash_screen_container"
-    app:layoutDescription="@xml/activity_splash_scene"
-    tools:context=".ui.activity.main.splash.SplashActivity">
-
-    <TextView
-        android:id="@+id/app_name_text"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:gravity="center_horizontal"
-        android:text="@string/text_title_target_mvd"
-        android:textColor="@android:color/black"
-        android:textStyle="bold" />
-
-    <ImageView
-        android:id="@+id/background_image"
-        android:layout_width="match_parent"
-        android:layout_height="@dimen/height_main_header"
-        android:layout_marginBottom="@dimen/margin_negative_50"
-        android:rotationX="180"
-        android:src="@drawable/bc_header_green_yellow"
-        app:layout_constraintBottom_toBottomOf="parent" />
-</androidx.constraintlayout.motion.widget.MotionLayout>
