+++ app/src/main/res/values/dimens.xml
@@ -9,7 +9,6 @@
     <dimen name="text_input_margin_top">15dp</dimen>
     <dimen name="sign_up_margin_horizontal">45dp</dimen>
     <dimen name="sign_up_button_height">48dp</dimen>
-    <dimen name="text_size_30">30sp</dimen>
     <dimen name="text_size_20">20sp</dimen>
     <dimen name="text_size_12">12sp</dimen>
     <dimen name="text_size_14">14sp</dimen>
