--- app/src/main/res/xml/activity_splash_scene.xml
@@ -1,80 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<MotionScene xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-
-    <Transition
-        android:id="@+id/transitionToEnd"
-        app:autoTransition="animateToEnd"
-        app:constraintSetEnd="@+id/end"
-        app:constraintSetStart="@id/start"
-        app:duration="1500" >
-        <KeyFrameSet>
-            <KeyAttribute
-                app:motionTarget="@+id/background_image"
-                app:framePosition="0"
-                android:scaleX="1.0" />
-            <KeyAttribute
-                app:motionTarget="@+id/background_image"
-                app:framePosition="0"
-                android:scaleY="1.0" />
-
-            <KeyAttribute
-                app:motionTarget="@+id/background_image"
-                app:framePosition="20"
-                android:scaleX="2.5" />
-            <KeyAttribute
-                app:motionTarget="@+id/background_image"
-                app:framePosition="20"
-                android:scaleY="2.5" />
-
-            <KeyAttribute
-                app:motionTarget="@+id/background_image"
-                app:framePosition="40"
-                android:scaleX="10.0" />
-            <KeyAttribute
-                app:motionTarget="@+id/background_image"
-                app:framePosition="40"
-                android:scaleY="10.0" />
-
-            <KeyAttribute
-                app:motionTarget="@+id/background_image"
-                app:framePosition="100"
-                android:scaleX="20.0" />
-            <KeyAttribute
-                app:motionTarget="@+id/background_image"
-                app:framePosition="100"
-                android:scaleY="20.0" />
-        </KeyFrameSet>
-    </Transition>
-
-    <ConstraintSet android:id="@+id/start">
-        <Constraint
-            android:id="@+id/app_name_text"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:gravity="center_horizontal"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
-            app:transitionEasing="accelerate"
-            app:layout_constraintVertical_bias="0.36">
-            <CustomAttribute
-                app:attributeName="textSize"
-                app:customFloatValue="30" />
-        </Constraint>
-    </ConstraintSet>
-
-    <ConstraintSet android:id="@+id/end">
-        <Constraint
-            android:id="@+id/app_name_text"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="@dimen/margin_big_45"
-            android:gravity="center_horizontal"
-            app:layout_constraintTop_toTopOf="parent">
-            <CustomAttribute
-                app:attributeName="textSize"
-                app:customFloatValue="20" />
-
-        </Constraint>
-    </ConstraintSet>
-</MotionScene>
